+++ src/main/java/core/prediction/TeamRatingPanel.java
@@ -14,13 +14,15 @@ import javax.swing.*;
 import java.awt.*;
 import java.awt.event.ItemEvent;
 import java.awt.event.ItemListener;
+import java.io.Serial;
 import java.util.ArrayList;
 import java.util.List;
 
 
 
 class TeamRatingPanel extends JPanel implements ItemListener {
-	private static final long serialVersionUID = -6238120571629957579L;
+	@Serial
+    private static final long serialVersionUID = -6238120571629957579L;
     //~ Instance fields ----------------------------------------------------------------------------
 	private final GridBagConstraints m_clConstraints;
     private List<RatingItem> levels;
@@ -78,12 +80,12 @@ class TeamRatingPanel extends JPanel implements ItemListener {
             tactIndex = 6;
         }
 
-        values[row][0].setSelectedIndex(tactIndex);
+        values[row][0].setSelectedIndex(Math.max(0, Math.min(tactIndex, tactics.size()-1)));
         values[row][0].addItemListener(this);
         taktikpanel.add(values[row][0]);
 
         values[row][1] = new JComboBox(levels.toArray());
-        values[row][1].setSelectedIndex(Math.min(team.getTacticLevel(), 19)); // limit tactic strength to divine
+        values[row][1].setSelectedIndex(Math.max(0,Math.min(team.getTacticLevel(), levels.size()-1)));
         values[row][1].addItemListener(this);
 
         if (team.getTacticType() == IMatchDetails.TAKTIK_NORMAL) {