+++ src/main/resources/release_notes.md
@@ -14,8 +14,10 @@
 
 * Calculation of the tsi sub by approximating the tsi formula (#235)
 
+### Team Analyzer
+* Fix illegal argument exception in team rating panel (#2155)
+* 
 ### Player Analysis
-
 * Fix error in rating compare table (#2153)
 
 ### Rating